requirements.txt
pymysql
django-material